Refactor: Move from widget duirectory to components directory
