+++ lib/components/templates/bottom_nav_layout.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import '../screens/home_page.dart';
-import '../screens/register_toilet_page.dart';
-import '../screens/profile_page.dart';
-import '../screens/notifications_page.dart';
+import '../../screens/home_page.dart';
+import '../../screens/register_toilet_page.dart';
+import '../../screens/profile_page.dart';
+import '../../screens/notifications_page.dart';
 
 class BottomNavLayout extends StatefulWidget {
   final int currentIndex;
